Move useFetchProductsIfNotLoaded to own file
