+++ client/my-sites/checkout/composite-checkout/hooks/use-fetch-products-if-not-loaded.ts
@@ -1,0 +1,27 @@
+/**
+ * External dependencies
+ */
+import { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import debugFactory from 'debug';
+
+/**
+ * Internal dependencies
+ */
+import { getProductsList, isProductsListFetching } from 'state/products-list/selectors';
+import { requestProductsList } from 'state/products-list/actions';
+
+const debug = debugFactory( 'calypso:composite-checkout:use-prepare-product-for-cart' );
+
+export default function useFetchProductsIfNotLoaded() {
+	const reduxDispatch = useDispatch();
+	const isFetchingProducts = useSelector( ( state ) => isProductsListFetching( state ) );
+	const products = useSelector( ( state ) => getProductsList( state ) );
+	useEffect( () => {
+		if ( ! isFetchingProducts && Object.keys( products || {} ).length < 1 ) {
+			debug( 'fetching products list' );
+			reduxDispatch( requestProductsList() );
+			return;
+		}
+	}, [ isFetchingProducts, products, reduxDispatch ] );
+}
